Validations and tests

+++ app/controllers/BookController.php
@@ -10,14 +10,25 @@ class BookController extends \BaseController {
 	public function index()
 	{
 		$books = Book::where('user_id', Auth::user()->id)->get();
- 
-		return Response::json(
-			array(
-				'error' => false,
-				'books' => $books->toArray()
-			),
-			200
-		);
+ 			
+ 		if($books) {
+			return Response::json(
+				array(
+					'error' => false,
+					'books' => $books->toArray()
+				),
+				200
+			);
+		}
+		else {
+			return Response::json(
+				array(
+					'error' 	=> true,
+					'message' 	=> 'This user has no books yet'
+				),
+				404
+			);
+		}
 	}
 
 	/**
@@ -46,34 +57,38 @@ class BookController extends \BaseController {
 
 
 		$rules = array(
-		        'title' => 'required',
-		        'description' => 'required'
+		        'title' 		=> 'required|min:3',
+		        'description' 	=> 'required|min:30'
 		    );
 
 
 		$validator = Validator::make(Input::all(), $rules);
 
-		if ($validator->fails())
-		{
-		   // Error pending...418 ?
+		if ($validator->fails()) {
+		   return Response::json(
+				array(
+					'error' 	=> true,
+					'message' 	=> 'The book creation failed'
+				),
+				404
+			);
 		}
 
-		else 
-		{
-			$book = new Book;
-			$book->title = Request::get('title');
-			$book->slug = uniqid();
-			$book->description = Request::get('description');
-			$book->user_id = Auth::user()->id;
-			$book->status = false;
+		else {
+			$book 				= new Book;
+			$book->title 		= Request::get('title');
+			$book->slug 		= uniqid();
+			$book->description 	= Request::get('description');
+			$book->user_id 		= Auth::user()->id;
+			$book->status 		= false;
 			$book->save();
 
 			return Response::json(
 				array(
 					'error' => false,
-					'book' => $book->toArray()
+					'book' 	=> $book->toArray()
 				),
-				200
+				201
 			);
 		}
 
@@ -88,11 +103,9 @@ class BookController extends \BaseController {
 	public function show($id)
 	{
 
-		$book = Book::where('id','=', $id)
-			->first();
+		$book = Book::where('id','=', $id)->first();
 
-		if($book)
-		{
+		if($book) {
 			return Response::json(
 				array(
 					'error' => false,
@@ -101,14 +114,13 @@ class BookController extends \BaseController {
 				200
 			);
 		}
-		else 
-		{
+		else {
 			return Response::json(
 				array(
-					'error' => true,
-					'message' => 'Book not found'
+					'error' 	=> true,
+					'message' 	=> 'Book not found'
 				),
-				400
+				404
 			);
 		}
 	}
@@ -132,29 +144,76 @@ class BookController extends \BaseController {
 	 */
 	public function update($id)
 	{
-		$modified = [];
-		$book = Book::where('user_id', Auth::user()->id)->find($id);
+		$modified 	= [];
+		$errors 	= [];
+		$book 		= Book::where('user_id', Auth::user()->id)->find($id);
 
-		if ( Request::get('title') ) {
-			$book->title = Request::get('title');
-			$modified['title'] = Request::get('title');
+		if($book) {
+
+			if ( Request::get('title') ) {
+
+				$validator = Validator::make(Input::all(), array('title' => 'required|min:3'));
+				if($validator->fails()) {
+					$errors['title'] = true;
+				}
+
+				else {
+					$book->title 		= Request::get('title');
+					$modified['title'] 	= Request::get('title');
+				}
+				
+			}
+
+			if ( Request::get('description') ) {
+
+				$validator = Validator::make(Input::all(), array('description' => 'required|min:30'));
+				if($validator->fails()) {
+					$errors['description'] = true;
+				}
+				else {
+				$book->description 			= Request::get('description');
+				$modified['description'] 	= Request::get('description');
+				}
+			}
+
+			// If validation doesn't pass
+
+			if(isset($errors['description']) OR isset($errors['title'])) {
+
+				return Response::json(
+					array(
+						'error' 	=> true,
+						'message' 	=> 'The title and/or the description isn\'t long enough'
+					),
+					204
+				);
+
+			}
+
+			// Else, we can update the book
+
+			else {
+				$book->save();
+
+				return Response::json(
+					array(
+						'error' 	=> false,
+						'modified' 	=> $modified,
+						'message' 	=> 'Book updated'
+					),
+					200
+				);
+			}
 		}
-
-		if ( Request::get('description') ) {
-			$book->description = Request::get('description');
-			$modified['description'] = Request::get('description');
+		else {
+			return Response::json(
+				array(
+					'error' 	=> true,
+					'message' 	=> 'Book doesn\'t exist'
+				),
+				404
+			);
 		}
-
-		$book->save();
-
-		return Response::json(
-			array(
-				'error' => false,
-				'modified' => $modified,
-				'message' => 'url updated'
-			),
-			200
-		);
 	}
 
 	/**
@@ -167,15 +226,27 @@ class BookController extends \BaseController {
 	{
 		$book = Book::where('user_id', Auth::user()->id)->find($id);
 
-		$book->delete();
+		if($book) {
+			$book->delete();
 
-		return Response::json(
-			array(
-				'error' => false,
-				'message' => 'book deleted'
-			),
-			200
-		);
+			return Response::json(
+				array(
+					'error' 	=> false,
+					'message'	 => 'book deleted'
+				),
+				200
+			);
+		}
+		else {
+			return Response::json(
+				array(
+					'error' => true,
+					'message' => 'Book doesn\'t exist'
+				),
+				404
+			);
+		}
+		
 	}
 
 }
